+++ src/page/account/FindPwPage/FindPwPage.jsx
@@ -1,15 +1,12 @@
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
 
+import { useFindPw } from '@/apis';
+
 import { Button, FetchLoadingOverlay, Icon, Input } from '@/shared/component';
 import { LOADING_MESSAGE } from '@/shared/constant';
 
-import { useFindPw } from '@/apis';
-
-import {
-  checkIfEntered,
-  checkMail,
-} from '@/pages/LoginPage/FindIdPage/inputCheck.js';
+import { checkIfEntered, checkMail } from '@/feature/account/lib';
 
 import styles from './FindPwPage.module.css';
