+++ src/page/account/SignUpFailure/SignUpFailurePage.module.css
