--- src/pages/LoginPage/SignUpPage/SignUpPageStages/index.js
@@ -1,3 +1,0 @@
-export { default as AccountInfoPage } from './AccountInfoPage';
-export { default as AuthorizationPage } from './AuthorizationPage';
-export { default as UserInfoPage } from './UserInfoPage';
