--- src/pages/LoginPage/index.js
@@ -1,10 +1,0 @@
-export { default as LoginPage } from './LoginPage/LoginPage';
-export { default as FindIdPage } from './FindIdPage/FindIdPage';
-export { default as FindPwPage } from './FindPwPage/FindPwPage';
-export { default as FoundIdPage } from './FoundIdPage/FoundIdPage';
-export { default as FoundPwPage } from './FoundPwPage/FoundPwPage';
-export { default as NotFoundIdPage } from './NotFoundIdPage/NotFoundIdPage';
-export { default as NotFoundPwPage } from './NotFoundPwPage/NotFoundPwPage';
-export { default as SignUpPage } from './SignUpPage/SignUpPage';
-export { default as SignUpSuccessPage } from './SignUpPage/SignUpSuccessPage/SignUpSuccessPage';
-export { default as SignUpFailurePage } from './SignUpPage/SignUpPageStages/SignUpFailure/SignUpFailurePage';
